# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Archive:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      created_by_id:
        type: integer
      description:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: integer
      metadata:
        description: JSONB field for custom metadata
        type: string
      status:
        $ref: '#/definitions/models.ArchiveStatus'
      tenant_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.ArchiveStatus:
    enum:
    - active
    - inactive
    - confidential
    type: string
    x-enum-varnames:
    - ArchiveStatusActive
    - ArchiveStatusInactive
    - ArchiveStatusConfidential
...
  models.Tenant:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      plan:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@kontena.io
    name: API Support
  description: Multi-tenant CRM API built with Golang and Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Kontena CRM API
  version: "1.0"
paths:
  /archives:
    get:
      consumes:
      - application/json
      description: Get all archives for the current tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Archive'
            type: array
      summary: Get all archives
      tags:
      - archives
    post:
      consumes:
      - application/json
      description: Create a new archive for the current tenant
      parameters:
      - description: Archive information
        in: body
        name: archive
        required: true
        schema:
          $ref: '#/definitions/models.Archive'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Archive'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an archive
      tags:
      - archives
  /archives/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an archive by ID for the current tenant
      parameters:
      - description: Archive ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an archive
      tags:
      - archives
    get:
      consumes:
      - application/json
      description: Get an archive by ID for the current tenant
      parameters:
      - description: Archive ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Archive'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an archive
      tags:
      - archives
    patch:
      consumes:
      - application/json
      description: Update an existing archive for the current tenant
      parameters:
      - description: Archive ID
        in: path
        name: id
        required: true
        type: integer
      - description: Archive information
        in: body
        name: archive
        required: true
        schema:
          $ref: '#/definitions/models.Archive'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Archive'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an archive
      tags:
      - archives
...
  /tenants:
    get:
      consumes:
      - application/json
      description: Get all tenants in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenant'
            type: array
      summary: Get all tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant
      parameters:
      - description: Tenant object
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a tenant
      tags:
      - tenants
  /tenants/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a tenant
      tags:
      - tenants
    get:
      consumes:
      - application/json
      description: Get a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a tenant
      tags:
      - tenants
    put:
      consumes:
      - application/json
      description: Update a tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: integer
      - description: Tenant object
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/models.Tenant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a tenant
      tags:
      - tenants
...
swagger: "2.0"